@@ -1,6 +1,6 @@
 {
   "name": "deadhash",
-  "version": "2.1.3",
+  "version": "2.2.0",
   "description": "File and text hash calculator",
   "homepage": "./",
   "private": true,
@@ -35,16 +35,18 @@
   },
   "main": "public/electron.js",
   "dependencies": {
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
+    "@emotion/react": "^11.4.1",
+    "@emotion/styled": "^11.3.0",
+    "@mui/icons-material": "^5.0.1",
+    "@mui/material": "^5.0.2",
     "crc": "^3.8.0",
     "cross-env": "^7.0.3",
     "electron-is-dev": "^2.0.0",
     "react": "^17.0.2",
     "react-contexify": "^5.0.0",
     "react-dom": "^17.0.2",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-router": "^5.2.1",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "^4.0.3"
   },
   "scripts": {
@@ -55,7 +57,7 @@
     "electron-build": "electron-builder",
     "release": "yarn react-build && electron-builder -wl --publish=always",
     "build": "yarn react-build && yarn electron-build -wl",
-    "start": "concurrently \"cross-env BROWSER=none yarn react-start\" \"wait-on http://localhost:3000 && electron .\""
+    "start": "yarn react-build && concurrently \"cross-env BROWSER=none yarn react-start\" \"wait-on http://localhost:3000 && electron .\""
   },
   "browserslist": {
     "production": [
@@ -70,15 +72,15 @@
     ]
   },
   "devDependencies": {
-    "concurrently": "^6.2.0",
-    "electron": "^13.1.1",
-    "electron-builder": "^22.10.5",
-    "eslint": "^7.28.0",
+    "concurrently": "^6.2.2",
+    "electron": "^15.0.0",
+    "electron-builder": "^22.11.7",
+    "eslint": "^7.32.0",
     "eslint-config-airbnb": "^18.2.1",
-    "eslint-plugin-import": "^2.23.4",
+    "eslint-plugin-import": "^2.24.2",
     "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.24.0",
+    "eslint-plugin-react": "^7.26.1",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "wait-on": "^5.3.0"
+    "wait-on": "^6.0.0"
   }
 }
