@@ -1,19 +1,24 @@
 import React from 'react';
-import Button from '@material-ui/core/Button';
-import ArrowLeftIcon from '@material-ui/icons/KeyboardArrowLeft';
+import Button from '@mui/material/Button';
+import ArrowLeftIcon from '@mui/icons-material/KeyboardArrowLeft';
+import { useHistory } from 'react-router-dom';
 
-const BackButton = ({ goBack }) => {
+const BackButton = () => {
+  const history = useHistory();
   /**
    * Call the goBack event
    * @param e The event argument
    */
   const back = (e) => {
     if (e) e.preventDefault();
-    if (goBack) goBack();
+    history.goBack();
   };
 
   return (
-    <Button color="primary" onClick={back}>
+    <Button
+      onClick={back}
+      size="small"
+    >
       <ArrowLeftIcon />
     </Button>
   );
