@@ -1,10 +1,10 @@
 import React from 'react';
-import Dialog from '@material-ui/core/Dialog';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
-import DialogActions from '@material-ui/core/DialogActions';
-import Button from '@material-ui/core/Button';
+import Dialog from '@mui/material/Dialog';
+import DialogTitle from '@mui/material/DialogTitle';
+import DialogContent from '@mui/material/DialogContent';
+import DialogContentText from '@mui/material/DialogContentText';
+import DialogActions from '@mui/material/DialogActions';
+import Button from '@mui/material/Button';
 
 const ConfirmationDialog = ({
   open, title, content, onAccept, onCancel, onClose, yes, no,
@@ -46,10 +46,10 @@ const ConfirmationDialog = ({
         </DialogContentText>
       </DialogContent>
       <DialogActions>
-        <Button onClick={cancel} color="primary">
+        <Button onClick={cancel}>
           {no}
         </Button>
-        <Button onClick={accept} color="primary" autoFocus>
+        <Button onClick={accept} autoFocus>
           {yes}
         </Button>
       </DialogActions>
