@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import Button from '@material-ui/core/Button';
+import Dialog from '@mui/material/Dialog';
+import DialogActions from '@mui/material/DialogActions';
+import DialogContent from '@mui/material/DialogContent';
+import DialogContentText from '@mui/material/DialogContentText';
+import DialogTitle from '@mui/material/DialogTitle';
+import Button from '@mui/material/Button';
 
 const UpdateDialog = ({
   downloadUrl, infoUrl, newVersion, updateAvailable, newVersionText, information, download, cancel,
@@ -50,10 +50,10 @@ const UpdateDialog = ({
         <Button onClick={handleClose}>
           {cancel}
         </Button>
-        <Button onClick={() => openInformation()} color="primary">
+        <Button onClick={openInformation}>
           {information}
         </Button>
-        <Button onClick={openDownload} color="primary" autoFocus>
+        <Button onClick={openDownload} autoFocus>
           {download}
         </Button>
       </DialogActions>
