@@ -1,39 +1,17 @@
 import React, { useContext, useEffect, useState } from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Container from '@material-ui/core/Container';
-import Grid from '@material-ui/core/Grid';
-import Paper from '@material-ui/core/Paper';
-import Button from '@material-ui/core/Button';
-import Typography from '@material-ui/core/Typography';
-import RefreshIcon from '@material-ui/icons/Refresh';
-import { useHistory } from 'react-router';
-import BackButton from '../../components/BackButton';
+import Container from '@mui/material/Container';
+import Grid from '@mui/material/Grid';
+import Button from '@mui/material/Button';
+import RefreshIcon from '@mui/icons-material/Refresh';
+import Card from '@mui/material/Card';
+import CardContent from '@mui/material/CardContent';
+import Typography from '@mui/material/Typography';
 import Updater from '../../utils/Updater';
 import UpdateDialog from '../../components/UpdateDialog';
 import AlertDialog from '../../components/AlertDialog';
 import { setActiveListItem } from '../../reducers/MainReducer/Actions';
 import { MainContext } from '../../contexts/MainContextProvider';
-
-const useStyles = makeStyles((theme) => ({
-  content: {
-    flexGrow: 1,
-    overflow: 'auto',
-  },
-  heroContent: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(4, 0, 2),
-  },
-  container: {
-    paddingTop: theme.spacing(2),
-    paddingBottom: theme.spacing(2),
-  },
-  paper: {
-    padding: theme.spacing(2),
-    display: 'flex',
-    overflow: 'auto',
-    flexDirection: 'column',
-  },
-}));
+import PageHeader from '../../components/PageHeader';
 
 const os = window.require('os');
 const { ipcRenderer } = window.require('electron');
@@ -53,9 +31,6 @@ const About = () => {
   const [errorMessage, setErrorMessage] = useState(null);
   const [loading, setLoading] = useState(false);
   const [update, setUpdate] = useState(null);
-
-  const classes = useStyles();
-  const history = useHistory();
 
   useEffect(() => {
     dispatch(setActiveListItem(5));
@@ -83,26 +58,15 @@ const About = () => {
       });
   };
 
-  /**
-   * Go back to the previous page
-   */
-  const goBack = () => {
-    history.goBack();
-  };
-
   return (
     <div>
-      <div className={classes.heroContent}>
-        <Container maxWidth="sm">
-          <Typography variant="h4" align="center" color="textPrimary" gutterBottom>
-            {language.about}
-          </Typography>
-          <Typography variant="h6" align="center" color="textSecondary" paragraph>
-            {language.aboutSubtitle}
-          </Typography>
-        </Container>
-      </div>
-      <main className={classes.content}>
+      <PageHeader title={language.about} subtitle={language.aboutSubtitle} backButton />
+      <main
+        style={{
+          flexGrow: 1,
+          overflow: 'auto',
+        }}
+      >
         {update && update.updateAvailable ? (
           <UpdateDialog
             downloadUrl={update.updateUrl}
@@ -129,52 +93,45 @@ const About = () => {
             ok={language.ok}
           />
         ) : null}
-        <Container maxWidth="lg" className={classes.container}>
-          <Typography component="h2" variant="h5" color="primary" gutterBottom>
-            <BackButton goBack={goBack} />
-            {language.appName}
-            {' '}
-            -
-            {language.about}
-          </Typography>
-          <Paper className={classes.paper}>
-            <div style={{ whiteSpace: 'pre-wrap' }}>
-              <p>
-                {language.aboutMessage.replace('{x}', appVersion)}
-              </p>
-            </div>
-
-            <Grid container spacing={2}>
-              <Grid item xs={12} md={6} lg={6}>
-                <Button
-                  target="_blank"
-                  style={{ width: '100%' }}
-                  href="http://codedead.com/Software/DeadHash/gpl.pdf"
-                  color="primary"
-                  variant="contained"
-                >
-                  {language.license}
-                </Button>
+        <Container
+          maxWidth="lg"
+          style={{ marginTop: 10 }}
+        >
+          <Card>
+            <CardContent>
+              <Grid container spacing={2}>
+                <Grid item xs={12} md={12} lg={12}>
+                  <Typography style={{ whiteSpace: 'pre-wrap' }}>
+                    {language.aboutMessage.replace('{x}', appVersion)}
+                  </Typography>
+                </Grid>
+                <Grid item xs={12} md={6} lg={6}>
+                  <Button
+                    target="_blank"
+                    fullWidth
+                    href="https://codedead.com/Software/DeadHash/gpl.pdf"
+                    variant="contained"
+                  >
+                    {language.license}
+                  </Button>
+                </Grid>
+                <Grid item xs={12} md={6} lg={6}>
+                  <Button
+                    target="_blank"
+                    fullWidth
+                    href="https://codedead.com"
+                    variant="contained"
+                  >
+                    {language.codedead}
+                  </Button>
+                </Grid>
               </Grid>
-              <Grid item xs={12} md={6} lg={6}>
-                <Button
-                  target="_blank"
-                  style={{ width: '100%' }}
-                  href="http://codedead.com"
-                  color="primary"
-                  variant="contained"
-                >
-                  {language.codedead}
-                </Button>
-              </Grid>
-            </Grid>
-          </Paper>
+            </CardContent>
+          </Card>
           <Button
-            className={classes.button}
-            color="primary"
-            onClick={() => checkForUpdates()}
+            onClick={checkForUpdates}
             disabled={loading}
-            style={{ marginTop: 5 }}
+            style={{ marginTop: 10 }}
           >
             <RefreshIcon />
             {language.checkForUpdates}
