@@ -1,34 +1,17 @@
 import React, { useContext, useEffect, useState } from 'react';
-import Container from '@material-ui/core/Container';
-import Typography from '@material-ui/core/Typography';
-import { makeStyles } from '@material-ui/core';
-import Grid from '@material-ui/core/Grid';
-import CardActionArea from '@material-ui/core/CardActionArea';
-import CardMedia from '@material-ui/core/CardMedia';
-import CardContent from '@material-ui/core/CardContent';
-import Card from '@material-ui/core/Card';
+import Container from '@mui/material/Container';
+import Typography from '@mui/material/Typography';
+import Grid from '@mui/material/Grid';
+import CardActionArea from '@mui/material/CardActionArea';
+import CardContent from '@mui/material/CardContent';
+import Card from '@mui/material/Card';
 import { useHistory } from 'react-router';
-import blank from '../../components/Theme/blank.png';
 import Updater from '../../utils/Updater';
 import UpdateDialog from '../../components/UpdateDialog';
 import AlertDialog from '../../components/AlertDialog';
 import { setActiveListItem, setUpdateChecked } from '../../reducers/MainReducer/Actions';
 import { MainContext } from '../../contexts/MainContextProvider';
-
-const useStyles = makeStyles((theme) => ({
-  heroContent: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(4, 0, 2),
-  },
-  content: {
-    flexGrow: 1,
-    overflow: 'auto',
-  },
-  container: {
-    paddingTop: theme.spacing(2),
-    paddingBottom: theme.spacing(2),
-  },
-}));
+import PageHeader from '../../components/PageHeader';
 
 const os = window.require('os');
 
@@ -43,7 +26,6 @@ const Home = () => {
   const [errorMessage, setErrorMessage] = useState(null);
   const [update, setUpdate] = useState(null);
 
-  const classes = useStyles();
   const history = useHistory();
 
   /**
@@ -95,17 +77,13 @@ const Home = () => {
 
   return (
     <div>
-      <div className={classes.heroContent}>
-        <Container maxWidth="sm">
-          <Typography variant="h4" align="center" color="textPrimary" gutterBottom>
-            {language.cryptography}
-          </Typography>
-          <Typography variant="h6" align="center" color="textSecondary" paragraph>
-            {language.cryptographySubtitle}
-          </Typography>
-        </Container>
-      </div>
-      <main className={classes.content}>
+      <PageHeader title={language.cryptography} subtitle={language.cryptographySubtitle} />
+      <main
+        style={{
+          flexGrow: 1,
+          overflow: 'auto',
+        }}
+      >
         {update && update.updateAvailable ? (
           <UpdateDialog
             downloadUrl={update.updateUrl}
@@ -125,15 +103,14 @@ const Home = () => {
             ok={language.ok}
           />
         ) : null}
-        <Container maxWidth="lg" className={classes.container}>
+        <Container
+          maxWidth="lg"
+          style={{ marginTop: 10 }}
+        >
           <Grid container spacing={2}>
             <Grid item xs={12} md={6} lg={6}>
               <Card>
                 <CardActionArea onClick={() => openFileHasher()}>
-                  <CardMedia
-                    title={language.file}
-                    image={blank}
-                  />
                   <CardContent>
                     <Typography gutterBottom variant="h5" component="h2">
                       {language.file}
@@ -148,10 +125,6 @@ const Home = () => {
             <Grid item xs={12} md={6} lg={6}>
               <Card>
                 <CardActionArea onClick={() => openTextHasher()}>
-                  <CardMedia
-                    title={language.file}
-                    image={blank}
-                  />
                   <CardContent>
                     <Typography gutterBottom variant="h5" component="h2">
                       {language.text}
