@@ -1,34 +1,34 @@
 import React, { useContext, useEffect, useState } from 'react';
-import Container from '@material-ui/core/Container';
-import Typography from '@material-ui/core/Typography';
-import {
-  FormLabel, makeStyles, Radio, RadioGroup,
-} from '@material-ui/core';
-import Grid from '@material-ui/core/Grid';
-import blue from '@material-ui/core/colors/blue';
-import purple from '@material-ui/core/colors/purple';
-import deepPurple from '@material-ui/core/colors/deepPurple';
-import lightBlue from '@material-ui/core/colors/lightBlue';
-import red from '@material-ui/core/colors/red';
-import green from '@material-ui/core/colors/green';
-import lightGreen from '@material-ui/core/colors/lightGreen';
-import grey from '@material-ui/core/colors/grey';
-import orange from '@material-ui/core/colors/orange';
-import deepOrange from '@material-ui/core/colors/deepOrange';
-import amber from '@material-ui/core/colors/amber';
-import brown from '@material-ui/core/colors/brown';
-import Paper from '@material-ui/core/Paper';
-import FormControlLabel from '@material-ui/core/FormControlLabel';
-import Checkbox from '@material-ui/core/Checkbox';
-import FormControl from '@material-ui/core/FormControl';
-import InputLabel from '@material-ui/core/InputLabel';
-import Select from '@material-ui/core/Select';
-import MenuItem from '@material-ui/core/MenuItem';
-import Button from '@material-ui/core/Button';
-import RefreshIcon from '@material-ui/icons/Refresh';
-import { useHistory } from 'react-router';
+import Container from '@mui/material/Container';
+import Typography from '@mui/material/Typography';
+import FormLabel from '@mui/material/FormLabel';
+import Radio from '@mui/material/Radio';
+import RadioGroup from '@mui/material/RadioGroup';
+import Grid from '@mui/material/Grid';
+import blue from '@mui/material/colors/blue';
+import purple from '@mui/material/colors/purple';
+import deepPurple from '@mui/material/colors/deepPurple';
+import lightBlue from '@mui/material/colors/lightBlue';
+import red from '@mui/material/colors/red';
+import green from '@mui/material/colors/green';
+import lightGreen from '@mui/material/colors/lightGreen';
+import grey from '@mui/material/colors/grey';
+import orange from '@mui/material/colors/orange';
+import deepOrange from '@mui/material/colors/deepOrange';
+import amber from '@mui/material/colors/amber';
+import brown from '@mui/material/colors/brown';
+import FormGroup from '@mui/material/FormGroup';
+import FormControlLabel from '@mui/material/FormControlLabel';
+import Checkbox from '@mui/material/Checkbox';
+import FormControl from '@mui/material/FormControl';
+import InputLabel from '@mui/material/InputLabel';
+import Select from '@mui/material/Select';
+import MenuItem from '@mui/material/MenuItem';
+import Button from '@mui/material/Button';
+import RefreshIcon from '@mui/icons-material/Refresh';
+import Card from '@mui/material/Card';
+import CardContent from '@mui/material/CardContent';
 import Updater from '../../utils/Updater';
-import BackButton from '../../components/BackButton';
 import UpdateDialog from '../../components/UpdateDialog';
 import AlertDialog from '../../components/AlertDialog';
 import ConfirmationDialog from '../../components/ConfirmationDialog';
@@ -65,31 +65,7 @@ import {
   setCrc24State,
   setCrc32State,
 } from '../../reducers/CryptoReducer/Actions';
-
-const useStyles = makeStyles((theme) => ({
-  content: {
-    flexGrow: 1,
-    overflow: 'auto',
-  },
-  heroContent: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(4, 0, 2),
-  },
-  container: {
-    paddingTop: theme.spacing(2),
-    paddingBottom: theme.spacing(2),
-  },
-  paper: {
-    padding: theme.spacing(2),
-    display: 'flex',
-    overflow: 'auto',
-    flexDirection: 'column',
-  },
-  button: {
-    marginTop: theme.spacing(1),
-    marginBottom: theme.spacing(1),
-  },
-}));
+import PageHeader from '../../components/PageHeader';
 
 const os = window.require('os');
 
@@ -120,9 +96,6 @@ const Settings = () => {
   const [loading, setLoading] = useState(false);
   const [update, setUpdate] = useState(null);
   const [confirmOpen, setConfirmOpen] = useState(false);
-
-  const classes = useStyles();
-  const history = useHistory();
 
   useEffect(() => {
     d1(setActiveListItem(3));
@@ -182,26 +155,15 @@ const Settings = () => {
     d1(setThemeStyle(event.target.value));
   };
 
-  /**
-   * Go back to the previous page
-   */
-  const goBack = () => {
-    history.goBack();
-  };
-
   return (
     <div>
-      <div className={classes.heroContent}>
-        <Container maxWidth="sm">
-          <Typography variant="h4" align="center" color="textPrimary" gutterBottom>
-            {language.settings}
-          </Typography>
-          <Typography variant="h6" align="center" color="textSecondary" paragraph>
-            {language.settingsSubtitle}
-          </Typography>
-        </Container>
-      </div>
-      <main className={classes.content}>
+      <PageHeader title={language.settings} subtitle={language.settingsSubtitle} backButton />
+      <main
+        style={{
+          flexGrow: 1,
+          overflow: 'auto',
+        }}
+      >
         {update && update.updateAvailable ? (
           <UpdateDialog
             downloadUrl={update.updateUrl}
@@ -228,115 +190,115 @@ const Settings = () => {
             ok={language.ok}
           />
         ) : null}
-        <Container maxWidth="lg" className={classes.container}>
+        <Container
+          maxWidth="lg"
+          style={{ marginTop: 10 }}
+        >
           <Grid container spacing={2}>
             <Grid item xs={12} md={12} lg={12}>
               <Typography component="h2" variant="h5" color="primary" gutterBottom>
-                <BackButton goBack={goBack} />
                 {language.general}
               </Typography>
-              <Paper className={classes.paper}>
-                <FormControlLabel
-                  control={(
-                    <Checkbox
-                      checked={autoUpdate}
-                      onChange={(e) => d1(setAutoUpdate(e.target.checked))}
-                      value="autoUpdate"
-                      color="primary"
+              <Card>
+                <CardContent>
+                  <FormGroup>
+                    <FormControlLabel
+                      control={(
+                        <Checkbox
+                          checked={autoUpdate}
+                          onChange={(e) => d1(setAutoUpdate(e.target.checked))}
+                          value="autoUpdate"
+                        />
+                      )}
+                      label={language.autoUpdate}
                     />
-                  )}
-                  label={language.autoUpdate}
-                />
-                <FormControlLabel
-                  control={(
-                    <Checkbox
-                      checked={canDragDrop}
-                      onChange={(e) => d1(setCanDragDrop(e.target.checked))}
-                      value="canDragDrop"
-                      color="primary"
+                    <FormControlLabel
+                      control={(
+                        <Checkbox
+                          checked={canDragDrop}
+                          onChange={(e) => d1(setCanDragDrop(e.target.checked))}
+                          value="canDragDrop"
+                        />
+                      )}
+                      label={language.dragAndDrop}
                     />
-                  )}
-                  label={language.dragAndDrop}
-                />
-                <FormControlLabel
-                  control={(
-                    <Checkbox
-                      checked={minimizeEnabled}
-                      onChange={(e) => d1(setMinimizeStatus(e.target.checked))}
-                      value="minimizeEnabled"
-                      color="primary"
+                    <FormControlLabel
+                      control={(
+                        <Checkbox
+                          checked={minimizeEnabled}
+                          onChange={(e) => d1(setMinimizeStatus(e.target.checked))}
+                          value="minimizeEnabled"
+                        />
+                      )}
+                      label={language.minimizeEnabled}
                     />
-                  )}
-                  label={language.minimizeEnabled}
-                />
-                <FormControlLabel
-                  control={(
-                    <Checkbox
-                      checked={maximizeEnabled}
-                      onChange={(e) => d1(setMaximizeStatus(e.target.checked))}
-                      value="maximizeEnabled"
-                      color="primary"
+                    <FormControlLabel
+                      control={(
+                        <Checkbox
+                          checked={maximizeEnabled}
+                          onChange={(e) => d1(setMaximizeStatus(e.target.checked))}
+                          value="maximizeEnabled"
+                        />
+                      )}
+                      label={language.maximizeEnabled}
                     />
-                  )}
-                  label={language.maximizeEnabled}
-                />
-                <FormControlLabel
-                  control={(
-                    <Checkbox
-                      checked={languageEnabled}
-                      onChange={(e) => d1(setLanguageButtonStatus(e.target.checked))}
-                      value="languageEnabled"
-                      color="primary"
+                    <FormControlLabel
+                      control={(
+                        <Checkbox
+                          checked={languageEnabled}
+                          onChange={(e) => d1(setLanguageButtonStatus(e.target.checked))}
+                          value="languageEnabled"
+                        />
+                      )}
+                      label={language.languageEnabled}
                     />
-                  )}
-                  label={language.languageEnabled}
-                />
-                <FormControlLabel
-                  control={(
-                    <Checkbox
-                      checked={themeToggleEnabled}
-                      onChange={(e) => d1(setThemeToggleStatus(e.target.checked))}
-                      value="themeToggleEnabled"
-                      color="primary"
+                    <FormControlLabel
+                      control={(
+                        <Checkbox
+                          checked={themeToggleEnabled}
+                          onChange={(e) => d1(setThemeToggleStatus(e.target.checked))}
+                          value="themeToggleEnabled"
+                        />
+                      )}
+                      label={language.themeToggleEnabled}
                     />
-                  )}
-                  label={language.themeToggleEnabled}
-                />
-                <FormControl variant="outlined" style={{ marginTop: 5 }}>
-                  <InputLabel id="language-label">{language.language}</InputLabel>
-                  <Select
-                    value={languageIndex}
-                    onChange={handleLanguageChange}
-                    id="language-simple"
-                    labelId="language-label"
-                  >
-                    <MenuItem value={0}>Deutsch</MenuItem>
-                    <MenuItem value={1}>English</MenuItem>
-                    <MenuItem value={2}>Español</MenuItem>
-                    <MenuItem value={3}>Français</MenuItem>
-                    <MenuItem value={4}>Italiano</MenuItem>
-                    <MenuItem value={5}>日本語</MenuItem>
-                    <MenuItem value={6}>Nederlands</MenuItem>
-                    <MenuItem value={7}>Português</MenuItem>
-                    <MenuItem value={8}>Pусский</MenuItem>
-                    <MenuItem value={9}>Türkçe</MenuItem>
-                  </Select>
-                </FormControl>
-              </Paper>
+                    <FormControl variant="outlined" style={{ marginTop: 5 }}>
+                      <InputLabel id="language-label">{language.language}</InputLabel>
+                      <Select
+                        value={languageIndex}
+                        onChange={handleLanguageChange}
+                        id="language-simple"
+                        labelId="language-label"
+                        label={language.language}
+                      >
+                        <MenuItem value={0}>Deutsch</MenuItem>
+                        <MenuItem value={1}>English</MenuItem>
+                        <MenuItem value={2}>Español</MenuItem>
+                        <MenuItem value={3}>Français</MenuItem>
+                        <MenuItem value={4}>Italiano</MenuItem>
+                        <MenuItem value={5}>日本語</MenuItem>
+                        <MenuItem value={6}>Nederlands</MenuItem>
+                        <MenuItem value={7}>Português</MenuItem>
+                        <MenuItem value={8}>Pусский</MenuItem>
+                        <MenuItem value={9}>Türkçe</MenuItem>
+                      </Select>
+                    </FormControl>
+                  </FormGroup>
+                </CardContent>
+              </Card>
             </Grid>
             <Grid item xs={12} md={12} lg={12}>
               <Typography component="h2" variant="h5" color="primary" gutterBottom>
                 {language.cryptography}
               </Typography>
-              <Paper className={classes.paper}>
-                <div>
+              <Card>
+                <CardContent>
                   <FormControlLabel
                     control={(
                       <Checkbox
                         checked={md4}
                         onChange={(e) => d2(setMd4State(e.target.checked))}
                         value="md4"
-                        color="primary"
                       />
                     )}
                     label={language.md4}
@@ -348,7 +310,6 @@ const Settings = () => {
                         checked={md5}
                         onChange={(e) => d2(setMd5State(e.target.checked))}
                         value="md5"
-                        color="primary"
                       />
                     )}
                     label={language.md5}
@@ -360,7 +321,6 @@ const Settings = () => {
                         checked={sha1}
                         onChange={(e) => d2(setSha1State(e.target.checked))}
                         value="sha1"
-                        color="primary"
                       />
                     )}
                     label={language.sha1}
@@ -372,7 +332,6 @@ const Settings = () => {
                         checked={sha224}
                         onChange={(e) => d2(setSha224State(e.target.checked))}
                         value="sha224"
-                        color="primary"
                       />
                     )}
                     label={language.sha224}
@@ -384,7 +343,6 @@ const Settings = () => {
                         checked={sha256}
                         onChange={(e) => d2(setSha256State(e.target.checked))}
                         value="sha256"
-                        color="primary"
                       />
                     )}
                     label={language.sha256}
@@ -396,7 +354,6 @@ const Settings = () => {
                         checked={sha384}
                         onChange={(e) => d2(setSha384State(e.target.checked))}
                         value="sha384"
-                        color="primary"
                       />
                     )}
                     label={language.sha384}
@@ -408,7 +365,6 @@ const Settings = () => {
                         checked={sha512}
                         onChange={(e) => d2(setSha512State(e.target.checked))}
                         value="sha512"
-                        color="primary"
                       />
                     )}
                     label={language.sha512}
@@ -420,27 +376,25 @@ const Settings = () => {
                         checked={ripemd160}
                         onChange={(e) => d2(setRipeMd160State(e.target.checked))}
                         value="ripemd160"
-                        color="primary"
                       />
                     )}
                     label={language.ripemd160}
                   />
-                </div>
-              </Paper>
+                </CardContent>
+              </Card>
             </Grid>
             <Grid item xs={12} md={12} lg={12}>
               <Typography component="h2" variant="h5" color="primary" gutterBottom>
                 {language.cyclicRedundancyCheck}
               </Typography>
-              <Paper className={classes.paper}>
-                <div>
+              <Card>
+                <CardContent>
                   <FormControlLabel
                     control={(
                       <Checkbox
                         checked={crc1}
                         onChange={(e) => d2(setCrc1State(e.target.checked))}
                         value="crc1"
-                        color="primary"
                       />
                     )}
                     label={language.crc1}
@@ -451,7 +405,6 @@ const Settings = () => {
                         checked={crc8}
                         onChange={(e) => d2(setCrc8State(e.target.checked))}
                         value="crc8"
-                        color="primary"
                       />
                     )}
                     label={language.crc8}
@@ -462,7 +415,6 @@ const Settings = () => {
                         checked={crc16}
                         onChange={(e) => d2(setCrc16State(e.target.checked))}
                         value="crc16"
-                        color="primary"
                       />
                     )}
                     label={language.crc16}
@@ -473,7 +425,6 @@ const Settings = () => {
                         checked={crc24}
                         onChange={(e) => d2(setCrc24State(e.target.checked))}
                         value="crc24"
-                        color="primary"
                       />
                     )}
                     label={language.crc24}
@@ -484,13 +435,12 @@ const Settings = () => {
                         checked={crc32}
                         onChange={(e) => d2(setCrc32State(e.target.checked))}
                         value="crc32"
-                        color="primary"
                       />
                     )}
                     label={language.crc32}
                   />
-                </div>
-              </Paper>
+                </CardContent>
+              </Card>
             </Grid>
             <Grid item xs={12} md={12} lg={12}>
               <Typography component="h2" variant="h5" color="primary" gutterBottom>
@@ -585,9 +535,9 @@ const Settings = () => {
               </GridList>
             </Grid>
             <Grid item xs={12} md={12} lg={12}>
-              <FormControl style={{ marginTop: 10 }}>
-                <FormLabel>{language.themeStyle}</FormLabel>
-                <RadioGroup value={themeStyle} onChange={changeThemeStyle}>
+              <FormControl style={{ marginTop: 10 }} component="fieldset">
+                <FormLabel component="legend">{language.themeStyle}</FormLabel>
+                <RadioGroup row value={themeStyle} onChange={changeThemeStyle}>
                   <FormControlLabel value="light" control={<Radio />} label={language.light} />
                   <FormControlLabel value="dark" control={<Radio />} label={language.dark} />
                 </RadioGroup>
@@ -595,8 +545,7 @@ const Settings = () => {
             </Grid>
           </Grid>
           <Button
-            className={classes.button}
-            color="primary"
+            style={{ marginTop: 10 }}
             onClick={() => checkForUpdates()}
             disabled={loading}
           >
@@ -604,11 +553,9 @@ const Settings = () => {
             {language.checkForUpdates}
           </Button>
           <Button
-            color="primary"
             variant="contained"
             onClick={() => setConfirmOpen(true)}
-            className={classes.button}
-            style={{ float: 'right' }}
+            style={{ marginTop: 10, float: 'right' }}
           >
             {language.reset}
           </Button>
